Issues #120 refactor: WorkManager 파일 정리
