+++ presentation/src/main/java/com/lighthouse/presentation/background/NotificationWorker.kt
@@ -1,12 +1,8 @@
-package com.lighthouse.presentation.notification
+package com.lighthouse.presentation.background
 
 import android.content.Context
 import androidx.hilt.work.HiltWorker
 import androidx.work.CoroutineWorker
-import androidx.work.ExistingPeriodicWorkPolicy
-import androidx.work.PeriodicWorkRequest
-import androidx.work.PeriodicWorkRequestBuilder
-import androidx.work.WorkManager
 import androidx.work.WorkerParameters
 import com.lighthouse.domain.model.DbResult
 import com.lighthouse.domain.model.Gifticon
@@ -14,8 +10,6 @@ import com.lighthouse.domain.usecase.GetGifticonsUseCase
 import com.lighthouse.presentation.util.TimeCalculator
 import dagger.assisted.Assisted
 import dagger.assisted.AssistedInject
-import dagger.hilt.android.qualifiers.ApplicationContext
-import java.util.concurrent.TimeUnit
 
 @HiltWorker
 class NotificationWorker @AssistedInject constructor(
@@ -46,19 +40,5 @@ class NotificationWorker @AssistedInject constructor(
                 createNotification(remain14Days, 14)
             }
         }
-    }
-}
-
-class NotificationWorkManager(@ApplicationContext context: Context) {
-    private val notificationWorkRequest: PeriodicWorkRequest =
-        PeriodicWorkRequestBuilder<NotificationWorker>(15, TimeUnit.MINUTES).build()
-
-    init {
-        WorkManager.getInstance(context)
-            .enqueueUniquePeriodicWork(WORK_NAME, ExistingPeriodicWorkPolicy.KEEP, notificationWorkRequest)
-    }
-
-    companion object {
-        private const val WORK_NAME = "notification"
     }
 }
